[#177] Move location permission warning/request to a more logical location (user accepts a schedule) and other fixes
