+++ app/src/main/java/com/example/bigowlapp/fragments/ScheduleFormFragment.java
@@ -174,8 +174,8 @@ public class ScheduleFormFragment extends Fragment
                         Toast.makeText(getActivity(), "ADDED LOCATION TRACKING", Toast.LENGTH_LONG).show();
                     })
                     .addOnFailureListener(e -> {
-                        // TODO: this runs when the user has location services off
-                        Exception exception = e;
+                        // TODO: this runs when the user has location services off or
+                        //       location permissions denied
                         Toast.makeText(getActivity(), "FAILURE!!!!! TO ADD LOCATION TRACKING", Toast.LENGTH_LONG).show();
                     });
         });
